\name{EpicopyData-package}
\alias{EpicopyData-package}
\alias{EpicopyData}
\docType{package}
\title{
\packageTitle{EpicopyData}
}
\description{
\packageDescription{EpicopyData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EpicopyData}
\packageIndices{EpicopyData}
Data package for Epicopy.
}
\author{
\packageAuthor{EpicopyData}

Maintainer: \packageMaintainer{EpicopyData}
}
